Code cleanup - moved some logic to new component
